+++ examples/cmd/vectorUpsert/main.go
@@ -31,10 +31,10 @@ func main() {
 		Vectors: []request.Vector{
 			{
 				ID:     "id3",
-				Values: []float32{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0},
+				Values: []float64{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0},
 				SparseValues: &request.VectorSparseValues{
 					Indices: []int64{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
-					Values:  []float32{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0},
+					Values:  []float64{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0},
 				},
 				Metadata: map[string]interface{}{
 					"key1": "value1",
